sync commit: Multiprocess nvbandwidth, rodinia, 1GB shm benchE benchmarking

+++ cpu/cpu-client-ivshmem.h
@@ -18,17 +18,17 @@
 typedef struct __ivshmem_area {
     size_t max_size;
     size_t avail_size;             
-    off_t avail_offset; // returns lowest un-written offset.
+    uint64_t avail_offset; // returns lowest un-written offset.
 } _ivshmem_area;
 
 // One per cuda process i.e.
 // one per flyt libcudart instance
 typedef struct __ivshmem_clnt_ctx {
-    int shm_enabled;
+    uint8_t shm_enabled;
     int pid;
     ivshmem_setup_desc svc_args;
-    off_t shm_proc_start; // shm_mmap + shm_proc_start: lowest VA of the mmaped shm of this process.
-    int shm_proc_size;
+    uint64_t shm_proc_start; // shm_mmap + shm_proc_start: lowest VA of the mmaped shm of this process.
+    uint64_t shm_proc_size;
     void *shm_mmap;
     char shm_be_path[SHM_BE_PATH_SZ];
 
@@ -49,8 +49,8 @@ ivshmem_setup_desc *_clnt_mgr_get_shm(int clnt_pid, int clientd_mq_id);
 uintptr_t shm_get_writeaddr_clnt(ivshmem_clnt_ctx *ctx);
 uintptr_t shm_get_readaddr_clnt(ivshmem_clnt_ctx *ctx);
 
-off_t shm_get_write_area_offset(size_t sz); // default: 0
-off_t shm_get_read_area_offset(size_t sz); // default: 0
+uint64_t shm_get_write_area_offset(size_t sz); // default: 0
+uint64_t shm_get_read_area_offset(size_t sz); // default: 0
 
 int check_shm_limits( _ivshmem_area *area, int size);
